YJDH-65: add browser tests for logout and update company data test
