+++ frontend/shared/.eslintignore
@@ -1,2 +1,3 @@
 node_modules
 jest.config.js
+.eslintrc.js
