+++ frontend/shared/browser-tests/shared.components.ts
@@ -15,7 +15,6 @@ export const getSharedComponents = (t: TestController) => {
     const expectations = {
       async isNotPresent({ timeout } = { timeout: 10000 }) {
         await t
-          // eslint-disable-next-line security/detect-non-literal-fs-filename
           .expect(selectors.spinner().exists)
           .notOk(await getErrorMessage(t), { timeout });
       },
